feat: add options to describe command
